+++ modules/components/src/main/java/module-info.java
@@ -14,6 +14,9 @@ module mfx.components {
 	exports io.github.palexdev.mfxcomponents.controls.buttons;
 	exports io.github.palexdev.mfxcomponents.controls.fab;
 
+	// Layout
+	exports io.github.palexdev.mfxcomponents.layout;
+
 	// Skins
 	exports io.github.palexdev.mfxcomponents.skins;